Add tab completion to terminal
Added tab completion to the terminal. Also added some new hidden
commands as "easer eggs". The input element now spans the full width of
the terminal to prevent weird cut-off text. And the min width of the
stdout has been locked to 435px to make sure no text is cut-off.