Refactor current pages
There have been a few minor issues with the current html/css. To make
things a bit more maintainable these have been reworked and everything
should look nice and proper again.

The projects section still only allows for a very limited amount of
projects to be displayed, but a fix for that is upcoming.